pyproject.toml
[project]
name = "slackdump"
version = "0.1.0"
description = "Extract structured notes from Slack and add to Notion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tiktoken",
    "ipykernel",
    "pydantic",
    "notion-client",
    "openai",
    "pyyaml",
    "backoff>=2.2.1",
    "rapidfuzz>=3.10.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.3",
]

[project.scripts]
extract-page-id-mapping = "src.extract_page_id_mapping:main"
